Task done
This task implements user authorization by Email and password, while checking for the correctness of entering both Email and password, additionally checking for re-entering the password and comparing the two fields. After successful data entry, the Email is saved in cookies.

After entering the data correctly, the user is redirected to a page with filling in personal information, here the verification of the name, surname, year of birth, gender, phone number and Skype is implemented. After clicking on the Save button, the entered information is checked for correctness, if all fields are filled in correctly, then all fields are saved in cookies.

At the end of the session on the second page, the user is again redirected to the registration page, after successful re-entry, all the data entered correctly will be pulled up automatically.

The duration of the preservation of all cookies is one hour.